Added changes from the original NASS-1051 branch updating the remember me workflow
